+++ src/app/layout.tsx
@@ -6,6 +6,8 @@ import SnackbarComponent from "../packages/components/snackbar-component";
 import { store, Provider } from "../redux";
 import type { RootState } from "../redux/store";
 import ExpensesHome from "./expenses/page";
+import ExpensesControls from "../components/expense-controls/expenses-controls";
+import { useState } from "react";
 
 function ChildRootLayout({
   children,
@@ -16,10 +18,19 @@ function ChildRootLayout({
     (state: RootState) => state.snackbar
   );
 
+  const [addExpense, setAddExpense] = useState<boolean>(false);
+  const [resetExpenses, setResetExpenses] = useState<boolean>(false);
+
   return (
     <html lang="en">
       <body>
         <Provider store={store}>
+          <ExpensesControls
+            addExpense={addExpense}
+            resetExpenses={resetExpenses}
+            setAddExpense={setAddExpense}
+            setResetExpenses={setResetExpenses}
+          />
           <ExpensesHome />
           <SnackbarComponent
             autoHideDuration={2000}
